back-end/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
Django==4.2.11
django-debug-toolbar==4.3.0
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.51.0
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
inflection==0.5.1
install==1.3.5
ipython==8.18.1
jedi==0.19.1
kiwisolver==1.4.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
parso==0.8.3
pillow==10.3.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
wsproto==1.2.0
zipp==3.18.1
